chore: update ingress tests for ktf v0.0.14
This makes adjustments for KTF v0.0.14, but also made
a couple other adjustments:

- HTTPS specific testing has its own file now
- status checking is performed on v1.Ingress
